requirements.txt
databricks-sql-connector>=3.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
openai>=1.30.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
yfinance>=0.2.40
